+++ tests/units/Glpi/Features/Dcbreadcrumb.php
@@ -63,10 +63,9 @@ class DCBreadcrumb extends \DbTestCase
      * @param $result
      * @dataProvider DCBreadcrumbProvider
      */
-    public function testDCBreadcrumbProvider($class, $result)
+    public function testClassUsesTrait($class, $result)
     {
-        $this->login();
-        $this->boolean(method_exists($class, 'getDcBreadcrumbSpecificValueToDisplay'))->isIdenticalTo($result);
+        $this->boolean(in_array(\Glpi\Features\DCBreadcrumb::class, class_uses($class, true)));
     }
 
 
@@ -74,6 +73,101 @@ class DCBreadcrumb extends \DbTestCase
      * Test DCBreadCrumb
      */
     public function testDCBreadcrumb()
+    {
+        $this->login();
+
+       //create DataCenter
+        $datacenter = new \Datacenter();
+        $datacenter_name = 'test datacenter';
+        $datacenter_id = $datacenter->add($datacenter_input = [
+            'name'         => $datacenter_name,
+        ]);
+        $this->checkInput($datacenter, $datacenter_id, $datacenter_input);
+
+       //create DCRoom
+        $DCroom = new \DCRoom();
+        $DCroom_name = 'test room';
+        $DCroom_id = $DCroom->add($DCroom_input = [
+            'name'           => $DCroom_name,
+            'vis_cols'       => 10,
+            'vis_rows'       => 10,
+            'datacenters_id' => $datacenter_id
+        ]);
+        $this->checkInput($DCroom, $DCroom_id, $DCroom_input);
+
+       //create Rack
+        $rack = new \Rack();
+        $rack_name = 'test rack';
+        $rack_id = $rack->add($rack_input = [
+            'name'         => $rack_name,
+            'entities_id'   => 0,
+            'dcrooms_id'   => $DCroom_id,
+            'number_units' => 42,
+            'position' => 62,
+        ]);
+        $this->checkInput($rack, $rack_id, $rack_input);
+
+       //load computer and add it to rack
+        $computer1 = getItemByTypeName('Computer', '_test_pc01');
+
+       //create Rack_Item
+        $Itemrack = new \Item_Rack();
+        $rack_position = 37;
+        $Itemrack_id = $Itemrack->add($Itemrack_input = [
+            'racks_id' => $rack_id,
+            'itemtype' => "Computer",
+            'items_id' => $computer1->getID(),
+            'position' => $rack_position,
+        ]);
+        $this->checkInput($Itemrack, $Itemrack_id, $Itemrack_input);
+
+        $DCBreadcrumb = null;
+        $this->when(
+            function () use (&$DCBreadcrumb, $computer1) {
+                $DCBreadcrumb = \Computer::getDcBreadcrumbSpecificValueToDisplay($computer1->getID());
+            }
+        )->error()
+           ->withType(E_USER_DEPRECATED)
+           ->withMessage('Called method is deprecated')
+           ->exists()  // deprecation on getDcBreadcrumbSpecificValueToDisplay
+           ->exists()  // deprecation on getDcBreadcrumb
+           ->exists()  // deprecation on isEnclosurePart
+           ->exists(); // deprecation on isRackPart
+
+        $this->string($DCBreadcrumb)->isIdenticalTo(
+            sprintf(
+                "<i class='ti ti-building-warehouse'></i> %s &gt; <i class='ti ti-building'></i> %s &gt; <i class='ti ti-server'></i> %s&nbsp;(U%d)",
+                $datacenter_name,
+                $DCroom_name,
+                $rack_name,
+                $rack_position
+            )
+        );
+
+       //load computer without Rack link
+        $computer2 = getItemByTypeName('Computer', '_test_pc02');
+
+        $DCBreadcrumb = null;
+        $this->when(
+            function () use (&$DCBreadcrumb, $computer2) {
+                $DCBreadcrumb = \Computer::getDcBreadcrumbSpecificValueToDisplay($computer2->getID());
+            }
+        )->error()
+           ->withType(E_USER_DEPRECATED)
+           ->withMessage('Called method is deprecated')
+           ->exists()  // deprecation on getDcBreadcrumbSpecificValueToDisplay
+           ->exists()  // deprecation on getDcBreadcrumb
+           ->exists()  // deprecation on isEnclosurePart
+           ->exists(); // deprecation on isRackPart
+
+        $this->string($DCBreadcrumb)->isIdenticalTo('&nbsp;');
+    }
+
+
+    /**
+     * Test DCBreadCrumb
+     */
+    public function testRenderDcBreadcrumb()
     {
         $this->login();
 
@@ -151,66 +245,67 @@ class DCBreadcrumb extends \DbTestCase
         ]);
         $this->checkInput($Itemrack, $Itemrack_id, $Itemrack_input);
 
-        $DCBreadcrumb = \Computer::getDcBreadcrumbSpecificValueToDisplay($computer1->getID(), true, false);
+        $DCBreadcrumb = \Computer::renderDcBreadcrumb($computer1->getID());
 
-        $expected = sprintf(
-            "<div class=\"row\">
-   
-      <div class=\"col-auto p-1\">
-         <i class='ti ti-building-warehouse'></i> %s
+        $location_dcroom_completename = htmlentities($location_dcroom_completename); // > separator will be encoded in result
+        $expected = <<<HTML
+    <div class="row">
+                            <div class="col-auto p-1">
+                <i class='ti ti-building-warehouse'></i> {$datacenter_name}
 
-                     <br>
-            <span class='p-0 float-left badge bg-blue-lt d-inline-block text-truncate' data-bs-toggle='tooltip' title='%s' style='max-width: 100px;'>
-               <i class=\"ti ti-map-pin\"></i>%s
-            </span>
-               </div>
-               <div class=\"col-auto p-1\">
-            >
-         </div>
-         
-      <div class=\"col-auto p-1\">
-         <i class='ti ti-building'></i> %s
+                
+                                    <br>
+                    <span class="p-0 float-left badge bg-blue-lt d-inline-block text-truncate"
+                          data-bs-toggle="tooltip"
+                          title="{$location_datacenter_name}"
+                          style="max-width: 100px;">
+                        <i class="ti ti-map-pin"></i>
+                                                                              {$location_datacenter_name}
+                                            </span>
+                                                </div>
+                            <div class="col-auto p-1">
+                    &gt;
+                </div>
+                                <div class="col-auto p-1">
+                <i class='ti ti-building'></i> {$DCroom_name}
 
-                     <br>
-            <span class='p-0 float-left badge bg-blue-lt d-inline-block text-truncate' data-bs-toggle='tooltip' title='%s' style='max-width: 100px;'>
-               <i class=\"ti ti-map-pin\"></i>%s
-            </span>
-               </div>
-               <div class=\"col-auto p-1\">
-            >
-         </div>
-         
-      <div class=\"col-auto p-1\">
-         <i class='ti ti-server'></i> %s
+                
+                                    <br>
+                    <span class="p-0 float-left badge bg-blue-lt d-inline-block text-truncate"
+                          data-bs-toggle="tooltip"
+                          title="{$location_dcroom_completename}"
+                          style="max-width: 100px;">
+                        <i class="ti ti-map-pin"></i>
+                                                                                 &gt;
+                                                      {$location_dcroom_name}
+                                            </span>
+                                                </div>
+                            <div class="col-auto p-1">
+                    &gt;
+                </div>
+                                <div class="col-auto p-1">
+                <i class='ti ti-server'></i> {$rack_name}
 
-                     <br>
-            <span class='p-0 float-left badge bg-blue-lt d-inline-block text-truncate' data-bs-toggle='tooltip' title='%s' style='max-width: 100px;'>
-               <i class=\"ti ti-map-pin\"></i>%s
-            </span>
-               </div>
-               <div class=\"col-auto pt-1 p-0\">
-            (U%d)
-         </div>
-         </div>
-",
-            $datacenter_name,
-            $location_datacenter_name,
-            $location_datacenter_name,
-            $DCroom_name,
-            htmlentities($location_dcroom_completename),
-            htmlentities('> ') . $location_dcroom_name, // > is added (DCbreadcrumb cleanLocationInformations()) because is a child location of datecenter location
-            $rack_name,
-            $location_rack_name,
-            $location_rack_name,
-            $rack_position
-        );
+                                    (U{$rack_position})
+                
+                                    <br>
+                    <span class="p-0 float-left badge bg-blue-lt d-inline-block text-truncate"
+                          data-bs-toggle="tooltip"
+                          title="{$location_rack_name}"
+                          style="max-width: 100px;">
+                        <i class="ti ti-map-pin"></i>
+                                                   {$location_rack_name}
+                                            </span>
+                                                </div>
+                        </div>
+
+HTML;
         $this->string($DCBreadcrumb)->isIdenticalTo($expected);
 
        //load computer without Rack link
         $computer2 = getItemByTypeName('Computer', '_test_pc02');
 
-        $DCBreadcrumb = \Computer::getDcBreadcrumbSpecificValueToDisplay($computer2->getID());
-
-        $this->boolean($DCBreadcrumb)->isFalse();
+        $DCBreadcrumb = \Computer::renderDcBreadcrumb($computer2->getID());
+        $this->string($DCBreadcrumb)->isEmpty();
     }
 }