+++ CHANGELOG.md
@@ -22,6 +22,10 @@ The present file will list all changes made to the project; according to the
 #### Changes
 
 #### Deprecated
+- `Glpi\Features\DCBreadcrumb::getDcBreadcrumb()`
+- `Glpi\Features\DCBreadcrumb::getDcBreadcrumbSpecificValueToDisplay()`
+- `Glpi\Features\DCBreadcrumb::isEnclosurePart()`
+- `Glpi\Features\DCBreadcrumb::isRackPart()`
 - `Glpi\Event::showList()`
 - `Html::displayAjaxMessageAfterRedirect()`
 - `Toolbox::seems_utf8()`